requirements.txt
# Core Dependencies
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Image Processing and Augmentation
Pillow>=10.0.0

# Data Handling
pandas>=2.0.0

# Jupyter Notebook (optional, for exploration)
jupyter>=1.0.0
ipykernel>=6.25.0

# Progress Bars and Utilities
tqdm>=4.66.0

# Optional: For better plotting
seaborn>=0.12.0

# Optional: For model optimization
tensorflow-model-optimization>=0.7.0

# Optional: For enhanced video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.0